backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
pydantic>=2.9.0
python-dotenv>=1.0.1
groq>=0.11.0
google-genai>=1.0.0
youtube-transcript-api>=0.6.2
playwright>=1.47.0
beautifulsoup4>=4.12.3
httpx>=0.27.0
tavily-python>=0.3.3
lxml>=5.3.0
